chore(admin): move add route to /admin
