+++ src/app/admin/admin.route.ts
@@ -1,21 +1,20 @@
 import { Routes } from '@angular/router';
 
 import { UserRouteAccessService } from '@xm-ngx/core/auth';
-import {
-    auditsRoute,
-    clientMgmtRoute,
-    docsRoute,
-    formPlaygroundRoute,
-    gatewayRoute,
-    healthRoute,
-    logsRoute,
-    maintenanceRoute,
-    metricsRoute,
-    rolesMatrixRoute,
-    rolesMgmtRoute,
-    translationRoute,
-    userMgmtRoute,
-} from './';
+import { auditsRoute } from './audits.route';
+import { clientMgmtRoute } from './client-management.route';
+import { docsRoute } from './docs.route';
+import { formPlaygroundRoute } from './form-playground.route';
+import { gatewayRoute } from './gateway.route';
+import { healthRoute } from './health.route';
+import { logsRoute } from './logs.route';
+import { maintenanceRoute } from './maintenance.route';
+import { metricsRoute } from './metrics.route';
+import { rolesMgmtRoute } from './roles-management.route';
+import { rolesMatrixRoute } from './roles-matrix.route';
+import { translationRoute } from './translation.route';
+import { userMgmtRoute } from './user-management.route';
+
 
 const ADMIN_ROUTES = [
     auditsRoute,
