+++ src/app/admin/docs.route.ts
@@ -1,6 +1,6 @@
 import { Route } from '@angular/router';
 
-import { JhiDocsComponent } from './docs.component';
+import { JhiDocsComponent } from './docs/docs.component';
 
 export const docsRoute: Route = {
     path: 'docs',
