+++ src/app/admin/gateway.route.ts
@@ -1,6 +1,6 @@
 import { Route } from '@angular/router';
 
-import { JhiGatewayComponent } from './gateway.component';
+import { JhiGatewayComponent } from './gateway/gateway.component';
 
 export const gatewayRoute: Route = {
     path: 'gateway',
