+++ src/app/admin/logs.route.ts
@@ -1,6 +1,6 @@
 import { Route } from '@angular/router';
 
-import { LogsComponent } from './logs.component';
+import { LogsComponent } from './logs/logs.component';
 
 export const logsRoute: Route = {
     path: 'logs',
